Create Owner

+++ src/controllers/videosController.js
@@ -1,4 +1,5 @@
 import videoModel from "../models/Video";
+import userModel from "../models/User";
 
 // Create Video Controllers
 export const recommend = async (req, res) => {
@@ -10,7 +11,10 @@ export const recommend = async (req, res) => {
 
 export const watch = async (req, res) => {
     const { id } = req.params;
-    const findByIdVideo = await videoModel.findById(id);
+    
+    // Find Video By Id From Video Model And Fill owner information
+    const findByIdVideo = await videoModel.findById(id).populate("owner");
+
     if(!findByIdVideo) {
         return res.render("watch", { pageTitle: "Video not Found." });
     }
@@ -19,26 +23,46 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const { user: { _id } } = req.session;
     const findByIdVideo = await videoModel.findById(id);
+
+    // Check the Video
     if(!findByIdVideo) {
         return res.status(404).render("404", { pageTitle: "Video not Found.", });
     }
+
+    // Check the Owner of video
+    if( String(findByIdVideo.owner) !== String(_id) ) {
+        return res.status(403).redirect("/");
+    }
+
     return res.render("editVideo", { pageTitle: `Edit: ${findByIdVideo.title}`, findByIdVideo });
 };
 
 //Modify Videos
 export const postEdit = async (req, res) => {
     const { id } = req.params;
+    const { user: { _id } } = req.session;
     const { title, description, hashtags } = req.body;
     const findByIdVideo = await videoModel.exists({_id:id});
+    
+    // Check the Video
     if(!findByIdVideo) {
         return res.status(404).render("404", { pageTitle: "Video not Found.", });
     }
+
+    // Check the Owner of video
+    if( String(findByIdVideo.owner) !== String(_id) ) {
+        return res.status(403).redirect("/");
+    }
+
+    // Update Video
     await videoModel.findByIdAndUpdate(id, {
         title, 
         description, 
         hashtags: videoModel.formatHashtags(hashtags),
     });
+
     return res.redirect(`/videos/${id}`);
 };
 
@@ -60,9 +84,24 @@ export const search = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const { user: { _id } } = req.session;
     //console.log(id);
+
+    const findByIdVideo = await videoModel.findById(id);
+    
+    // Check the Video
+    if(!findByIdVideo) {
+        return res.status(404).render("404", { pageTitle: "Video not Found.", });
+    }
+
+    // Check the Owner of video
+    if( String(findByIdVideo.owner) !== String(_id) ) {
+        return res.status(403).redirect("/");
+    }
+
     // Delete Video
     await videoModel.findByIdAndDelete(id);
+
     return res.redirect("/");
 };
 
@@ -72,18 +111,28 @@ export const getUpload = (req, res) => {
 
 //Add Video 
 export const postUpload = async (req, res) => {
+    const { user: { _id } } = req.session;
     const { path: fileUrl } = req.file;
     const { videoTitle, description, hashtags } = req.body;
+    
     // Update On Database
     try {
         const newVideo = await videoModel.create({
             title: videoTitle,
             description,
             fileUrl,
+            owner: _id,
             hashtags: videoModel.formatHashtags(hashtags),
         });
-        //console.log(newVideo);
+
+        const user = await userModel.findById(_id);
+        user.videos.unshift(newVideo._id);
+        user.save();
+
+        // console.log(newVideo, user);
+
         return res.redirect("/");
+
     } catch(error){
         // Call Upload Page and Add Error Message
         return res.status(400).render("Upload", {
